cran2deb: rebuild the source tarball each time removing file exec bit.
The subdirectory of an R package is 'pkgname'.  Debian typically has
'pkgname-upstreamversion'.  Hence the tarball is rebuilt after renaming
the 'pkgname' directory appropriately.

Whilst this is all going on, every file in the R package has its
executable bit removed.  There would appear to be little correct need
for it at the moment (this will change when #! is handled better) and
this gets rid of some inappropriately executable files.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/cran2deb@23 edb9625f-4e0d-4859-8d74-9fd3b1da38cb