cran2deb: put base_pkgs into the cache.  generate cache so that R is happy. use separate base.tgz for cran2deb.
base_pkgs is the list of all packages that are provided in the basic
install of R.  It is found by listing all installed packages in the
pbuilder environment.

Previously the cache of availabile packages lived in sysdata.rda.
Unfortunately it appears that R does not like it when sysdata.rda is
updated after package installation (I think this is something to do with
lazy loading, but disabling this did not seem to help).  Instead the
cache is maintained separated in the data/ package directory.

pbuilder now uses base-cran2deb.tgz for the cran2deb pbuilder
environment -- this should help keep cran2deb from interfering from
other uses of pbuilder.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/cran2deb@54 edb9625f-4e0d-4859-8d74-9fd3b1da38cb