sysreq: parse SystemRequirements and check each one against database, after preprocessing.
``Writing R extensions'' does not actually specify a format for
SystemRequirements.  Fortunately most package authors seem to use
a similar notation as that of Depends/Imports, but also include URLs
and ad-hoc information.

This is enough to make the R package Matrix depend correctly on 'make'
and likely more after a bit of leg work.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/cran2deb@60 edb9625f-4e0d-4859-8d74-9fd3b1da38cb