license: split out some of the heuristics for reducing poorly formed license fields.
in particular, do not remove numbers in just one step. it seems common
for people to use:
License: X11 (http://blahblah/dfgdfg/)
which will be reduce to just 'x' if these steps are not split.

Also separate license hashing into getting the license text and then
separately hashing via digest. (Useful for for the 'view' command about to
be added to cran2deb license.)

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/cran2deb@76 edb9625f-4e0d-4859-8d74-9fd3b1da38cb