exec: drop bioc, more verbose version mismatch error, switch hoc to awk (more portable)
some packages are in bioc and cran (e.g., graph). this confused cran2deb
as One True Package is expected. to solve this, we could filter for
common tables and take the highest version number. but who's to say
these aren't actually distinct packages? so for now, assume all cran
dependencies are satisfiable in cran (they aren't).

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/cran2deb@96 edb9625f-4e0d-4859-8d74-9fd3b1da38cb