use NULL instead of NA. include git revision and date stamp in build log. generate changelog.
NA has length 1 and conveys perhaps-ok information. NULL was often more
appropriate.

a configure script edits R/zzz.R to add a new global 'git_revision'
prior to R building the package. This assumes that the current working
directory is in a git repository and will fail if it is not.

changelog generated according to Debian guidelines---latest builds to
the top. Included: the git revision of cran2deb, the time/date, and the
database version.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/cran2deb@111 edb9625f-4e0d-4859-8d74-9fd3b1da38cb