The way that the license hash_sha1 is computed in exec/license prduces the exact same results as does the UNIX tool sha1sum. This is a very decent thing and would be accepted by me as the reference of a license hash. Any other hash should be additive to that first one.
The routine producing the license text from files
with the 'file LICENSE' did some changes to the
license text prior to forming the hash sum. This
was certainly meant well, but it also prohibited
the "cran2deb license hash" to work together with
"cran2deb build" in exceptional cases when there
are newlines too much or the when some text of
the license had a blank at the end of the line,
and donno when else.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/cran2deb@358 edb9625f-4e0d-4859-8d74-9fd3b1da38cb